Use numpy 128 bits float
(cherry picked from commit fe913ee6154bc55d805992ab9a834f6c1f7c27dc)

+++ dataset/logic_circuit.py
@@ -8,7 +8,7 @@ import numpy as np
 # data 0,1番目が学習データ 2番目が答え
 def dset(d_name, num):
     if d_name == "or":
-        data = np.zeros((4 * num, 3))
+        data = np.zeros((4 * num, 3), dtype=np.float128)
         for i in range(num):
             data[4 * i] = [0, 0, 0]
             data[4*i + 1] = [0, 1, 1]
@@ -16,7 +16,7 @@ def dset(d_name, num):
             data[4*i + 3] = [1, 1, 1]
         return data
     elif d_name == "and":
-        data = np.zeros((4 * num, 3))
+        data = np.zeros((4 * num, 3), dtype=np.float128)
         for i in range(num):
             data[4 * i] = [0, 0, 0]
             data[4*i + 1] = [0, 1, 0]
@@ -24,7 +24,7 @@ def dset(d_name, num):
             data[4*i + 3] = [1, 1, 1]
         return data
     elif d_name == "nand":
-        data = np.zeros((4 * num, 3))
+        data = np.zeros((4 * num, 3), dtype=np.float128)
         for i in range(num):
             data[4 * i] = [0, 0, 1]
             data[4*i + 1] = [0, 1, 1]
@@ -32,7 +32,7 @@ def dset(d_name, num):
             data[4*i + 3] = [1, 1, 0]
         return data
     elif d_name == "xor":
-        data = np.zeros((4 * num, 3))
+        data = np.zeros((4 * num, 3), dtype=np.float128)
         for i in range(num):
             data[4 * i] = [0, 0, 1]
             data[4*i + 1] = [0, 1, 0]
@@ -40,7 +40,7 @@ def dset(d_name, num):
             data[4*i + 3] = [1, 1, 1]
         return data
     elif d_name == "w_not":
-        data = np.zeros((4 * num, 4))
+        data = np.zeros((4 * num, 4), dtype=np.float128)
         for i in range(num):
             data[4 * i] = [0, 0, 1, 1]
             data[4*i + 1] = [0, 1, 1, 0]
